Rename ahoy to internal
