+++ VideoApp/VideoApp/Stores/Auth/Internal/FirebaseAuthStore.swift
@@ -48,7 +48,7 @@ class FirebaseAuthStore: NSObject, FirebaseAuthStoreWriting {
     }
 
     func signIn(userIdentity: String, passcode: String, completion: @escaping (AuthError?) -> Void) {
-        print("Passcode sign in not supported by Firebase auth.")
+        fatalError("Passcode sign in not supported by Firebase auth.")
     }
 
     func signOut() {
